+++ app/core/+mag/+app/+control/HK.m
@@ -29,7 +29,7 @@ classdef HK < mag.app.Control & mag.app.mixin.StartEndDate
             this.Layout = this.createDefaultGridLayout(parent);
 
             % Start and end dates.
-            this.addStartEndDateButtons(this.Layout, Limits = this.Model.TimeRange);
+            this.addStartEndDateButtons(this.Layout, Limits = this.Model.HKTimeRange);
         end
 
         function supported = isSupported(~, results)