+++ app/mission/imap/+mag/+app/+imap/+control/Timestamp.m
@@ -16,7 +16,7 @@ classdef Timestamp < mag.app.Control & mag.app.mixin.StartEndDate
             this.Layout = this.createDefaultGridLayout(parent);
 
             % Start and end dates.
-            this.addStartEndDateButtons(this.Layout, Limits = this.Model.TimeRange);
+            this.addStartEndDateButtons(this.Layout, Limits = this.Model.ScienceTimeRange);
         end
 
         function supported = isSupported(~, results)