+++ app/mission/imap/+mag/+app/+imap/AnalysisManager.m
@@ -192,6 +192,8 @@ classdef AnalysisManager < mag.app.manage.AnalysisManager
             if ~isequal(location, 0)
                 this.LocationEditField.Value = location;
             end
+
+            this.LocationEditField.focus();
         end
     end
 end