+++ resources/release-notes.md
@@ -4,7 +4,11 @@
   ```
   settings().MAG.Toolbox.Sound
   ```
+- (All) Fix issue with `datetime` range slider showing the wrong time range for HK plots
+- (All) Fix issue with incompatible timezones in `datetime` slider and `datetime` range slider
+- (All) Force app to come back to focus after choosing folder in "Analysis" tab
 
 ## Software
 
+- (All) Add `Orientation` to `mag.graphics.chart.Line` to specify label orientation
 - (Bart) Fix issue with loading only one Bartington sensor data